02_FederatedLearning_ChameleonCloud/web_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.0.1
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
